pyproject.toml
[tool.poetry]
name = "repl_python3_beta_smolPixels"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
pydantic = "^1.8.2"
replit = "^3.1.0"
fastapi-limiter = "^0.1.4"
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
